Recomposition of specimen entries.

+++ Unite.Data.Context/Repositories/DonorsRepository.cs
@@ -1,6 +1,5 @@
 using Microsoft.EntityFrameworkCore;
-using Unite.Data.Entities.Donors;
-using Unite.Data.Entities.Genome.Variants;
+using Unite.Data.Entities.Genome.Analysis.Dna;
 using Unite.Data.Entities.Images;
 using Unite.Data.Entities.Images.Enums;
 using Unite.Data.Entities.Specimens;
@@ -55,7 +54,7 @@ public class DonorsRepository : Repository
         return await _specimensRepository.GetRelatedGenes(specimens);
     }
 
-    public async Task<long[]> GetRelatedVariants<TV>(IEnumerable<int> ids)
+    public async Task<int[]> GetRelatedVariants<TV>(IEnumerable<int> ids)
         where TV : Variant
     {
         var specimens = await GetRelatedSpecimens(ids);
